refactor(app): move root component to ./src
