+++ index.ts
@@ -1,6 +1,6 @@
 import { AppRegistry } from 'react-native';
 
-import App from './App';
-import { name as appName } from './app.json';
+import App from './src/App';
+import { name as appName } from './src/app.json';
 
 AppRegistry.registerComponent(appName, () => App);
