+++ src/App.spec.tsx
@@ -6,7 +6,7 @@ import 'react-native';
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import App from '../App';
+import App from './App';
 
 // Note: test renderer must be required after react-native.
